Further work on EntityData and Character Database TODO: Fix regression with Orientation
