+++ Projects/CoX/Common/GameData/chardata_definitions.h
@@ -10,8 +10,8 @@ struct CharacterData
     uint32_t    m_experience_debt   = 0;
     uint32_t    m_experience_patrol = 0; // planned future use
     uint32_t    m_influence         = 0;
-    bool        m_has_titles        = false;
-    bool        m_has_the_prefix    = false;
+    bool        m_has_titles;
+    bool        m_has_the_prefix;
     QString     m_titles[3];             // Generic, Origin, Special
     QString     m_battle_cry;
     QString     m_character_description;
