+++ Projects/CoX/Data/settings.cfg
@@ -1,3 +1,27 @@
+##############################################################
+#    SEGS configuration file.
+#
+#    listen_addr values below should contain the IP the
+#      clients will connect to.
+#
+#    location_addr values below should contain the IP the
+#      clients will receive data from.
+#
+#    Both values are set to 127.0.0.1 by default but should
+#      be set to your local IP address on the network
+#      for example: 10.0.0.2
+#
+#    Default ports are listed below:
+#      AccountDatabase db_port:		5432
+#      CharacterDatabase db_port:	5432
+#      AuthServer listen_addr:		2106
+#      GameServer listen_addr:		7002
+#      GameServer location_addr:	7002
+#      MapServer listen_addr:		7003
+#      MapServer location_addr:		7003
+#
+##############################################################
+
 [AdminServer]
 AccountDatabase\db_driver   = QSQLITE
 AccountDatabase\db_host     = 127.0.0.1
@@ -12,16 +36,16 @@ CharacterDatabase\db_name   = segs_game
 CharacterDatabase\db_user   = segsadmin
 CharacterDatabase\db_pass   = segs123
 
-[AuthServer] 
-listen_addr       = this_is_the_ip_the_clients_will_connect_to:2106
+[AuthServer]
+listen_addr       = 127.0.0.1:2106
 
 [GameServer] 
 server_name       = first server
-listen_addr       = this_is_the_ip_the_clients_will_connect_to:7002
-location_addr     = this_is_the_ip_the_clients_will_receive_from_server:7002
+listen_addr       = 127.0.0.1:7002
+location_addr     = 127.0.0.1:7002
 max_players       = 200
 max_account_slots = 8
 
 [MapServer]  
-listen_addr       = this_is_the_ip_the_clients_will_connect_to:7003
-location_addr     = this_is_the_ip_the_clients_will_receive_from_server:7003
+listen_addr       = 127.0.0.1:7003
+location_addr     = 127.0.0.1:7003
