+++ Projects/CoX/Servers/MapServer/Events/ChatMessage.cpp
@@ -4,7 +4,7 @@
 #include "Events/InputState.h"
 #include "Entity.h"
 #include "MapClient.h"
-#include "PlayerMethods.h"
+#include "Servers/MapServer/DataHelpers.h"
 
 #include <cmath>
